// apps/sage-sync/src/lib/sage/types.ts
type SageRecord = {
  RECORD_MODIFY_DATE: string;
  RECORD_CREATE_DATE: string;
  RECORD_DELETED: 0 | 1;
};
// ...
export type STOCK_TRAN = SageRecord & {
  TRAN_NUMBER: number;
  STOCK_CODE: string;
  TYPE: string;
  DATE: string;
  REFERENCE: string;
  REFERENCE_NUMERIC: number;
  DETAILS: string;
  QUANTITY: number;
  COST_PRICE: number;
  SALES_PRICE: number;
  QTY_USED: number;
  ISP_REFERENCE: number;
  ARTEFACT_TYPE: number;
  GDN_NUMBER: number;
  GRN_NUMBER: number;
};